Separate the nginx port from the Errbit port
If you're proxying to nginx rather than connecting to it directly then
the port that nginx listens on may well be different to the port that
Errbit should advertise. In our case, we have Apache accepting SSL
connections that proxy to nginx without SSL. It would be nice if the
cookbook could configure nginx for SSL but I'll leave that to someone
else. :)